requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
bcrypt==4.0.1
casbin==1.41.0
casbin_sqlalchemy_adapter==1.4.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
decorator==5.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jwt==1.3.1
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-jose==3.4.0
rsa==4.9
setuptools==75.8.0
simpleeval==1.0.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.46.1
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.34.0
wcwidth==0.2.13
wheel==0.45.1
